fixes geometry refinement
