+++ simtbx/modeling/predictions.py
@@ -198,7 +198,7 @@ def label_weak_spots_for_integration(fraction, predictions, num_res_bins=10):
     res_sort = np.sort(res)
     res_bins = [rb[0]-1e-6 for rb in np.array_split( res_sort, num_res_bins)] + [res_sort[-1]+1e-6]
     res_bin_assigments = np.digitize(res, res_bins)
-    is_weak_but_integratable = np.zeros(len(predictions)).astype(np.bool)
+    is_weak_but_integratable = np.zeros(len(predictions)).astype(bool)
     for i_res in range(1, num_res_bins+1):
         # grab weak spots in this res bin
         is_weak_and_in_bin = logi_and(res_bin_assigments == i_res, predictions["is_weak"])
